% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_lat_lon.R
\name{append_lat_lon}
\alias{append_lat_lon}
\title{Get 'centroid' coordinates in lat and lon}
\usage{
append_lat_lon(shp)
}
\arguments{
\item{shp}{The sf object}
}
\value{
A shape file with lat lon 'centroid' coordinates appended as columns before the geometries
}
\description{
This function extracts the centroid coordinates from the geometries of a shape file object and appends them as columns
}
\examples{
# shp_adm0 <- readRDS("analysis/data_derived/sf_admin0_africa.rds")
# shp_adm0 <- shp_adm0 \%>\% append_lat_lon()
}
