% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{determine_snp_and_grid_blocks_for_output}
\alias{determine_snp_and_grid_blocks_for_output}
\title{Mapping between output block and grid point assignments for each SNP}
\usage{
determine_snp_and_grid_blocks_for_output(
  grid,
  start_and_end_minus_buffer,
  outputSNPBlockSize = 1000
)
}
\arguments{
\item{grid:}{a positive, including 0, integer vector with length N SNPs.  Each
element i is the grid point that SNP i is assigned.}

\item{start_and_end_minus_buffer:}{a length 2 vector of positive integers.  The
elements 1:start_and_end_minus_buffer[1] and N-start_and_end_minus_buffer[2]:N
are considered buffer regions, with N being the number of SNPs.}

\item{outputSNPBlockSize:}{the number of SNPs that are output by the step
where SNP genotypes are computed.}
}
\value{
to_out: (number of output block rows, 4 columns) array that specifies
 the SNP indexes and grid point indexes per output block.
}
\description{
Grid points are used to reduce the number of loci that the 
 HMM calculations need to process.  At a given grid point, a set of nearby
 SNPs are assigned to that grid point.  Any one SNP must belong to a single
 grid point.  These grid points need to be compatible with the SNP output
 block size, that is, the approximate number of SNPs to apply the HMM for
 imputation.  For any one SNP output block, there must be at least two grid
 points and any one grid point must belong to a single output block.  This
 function applies these constraints to determine how grid points are mapped
 to output blocks.  Note, a corollary of these constraints is that any one
 grid point is mapped to a single outputblock.
}
