% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_hd_to_cube}
\alias{extract_hd_to_cube}
\title{Extract ancestral haplotype dosage to RData cube}
\usage{
extract_hd_to_cube(
  vcf_file,
  ref,
  bcftools = "bcftools",
  gatk_jar =
    "/data/smew1/rdavies/stitch_richard_paper/bin/gatk/GenomeAnalysisTK-3.8-1-0-gf15c1c3ef/GenomeAnalysisTK.jar",
  samples = NULL,
  pos = NULL,
  ram = "-Xmx4g",
  field = "HD",
  verbose = TRUE
)
}
\arguments{
\item{vcf_file}{path to VCF}

\item{ref}{path to reference fasta (required by GATK)}

\item{bcftools}{path to vcftools (or just "bcftools" if in path)}

\item{gatk_jar}{path to GATK jar}

\item{samples}{vector of sample names (or NULL for all)}

\item{pos}{pos matrix, a matrix with at least two columns where the first two columns are chrom and 1-based physical position, respectively. specifies which SNPs to extract (or NULL for all)}

\item{field}{What to get from the VCF. Default HD for haplotype dosages}
}
\value{
A cube with dimensions of SNPs x samples x ancestral haplotypes
}
\description{
Extract ancestral haplotype dosage to RData cube
}
\author{
Robert Davies
}
