% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_alignment_contig_length}
\alias{get_alignment_contig_length}
\title{Read a contig length from a BAM or CRAM header}
\usage{
get_alignment_contig_length(file, contig)
}
\arguments{
\item{file}{Path to a BAM or CRAM file.}

\item{contig}{Contig name as recorded in the `SN` header tag.}
}
\value{
The contig length as an integer.
}
\description{
Read a contig length from a BAM or CRAM header
}
