% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_covar_data.R
\name{get_map_covar_data}
\alias{get_map_covar_data}
\title{Title: Get map and covariate data in long-format}
\usage{
get_map_covar_data(covars_obj)
}
\arguments{
\item{covars_obj}{Description of the parameter 'x'.}
}
\value{
A sf object with covariate data in long format from all covar list
}
\description{
This unlists the covariate data frames and returns a sf object with the covariate data in long format.
It expects an object with a list of \code{covars} covariate data frames (eg by year)
and the \code{map} shp file from \code{scrape}. Bespoke function to wrangle \code{final_covariates.rds} from \code{scrape}.
}
\examples{
# Example of how to use the function
# covars <- readRDS("analysis/data_raw/final_covariates.rds")
# map_data <- get_map_covar_data(covars)
}
