% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pipe_source_name.R
\name{get_pipe_source_name}
\alias{get_pipe_source_name}
\title{get_pipe_source_name (non-exported)}
\usage{
get_pipe_source_name(df)
}
\arguments{
\item{data;}{data.frame-like object}
}
\description{
Get the name from the object at the front of a pipe.
}
\examples{

my_dt <- data.table(letters=letters[1:5])
fun <- function(data) {
  # browser()
  get_orig_name()
  # data
}


my_dt \%>\% get_pipe_source_name()
my_dt \%>\% fun()

# Not Exported
}
\references{
\href{https://stackoverflow.com/a/58540131/1504321}{Stackoverflow}
}
