% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic-map.R
\name{make_genetic_map_file_from_vcf}
\alias{make_genetic_map_file_from_vcf}
\title{Create a genetic map file from VCF assuming uniform recombination rate}
\usage{
make_genetic_map_file_from_vcf(outfile, vcffile, region = "", expRate = 0.5)
}
\arguments{
\item{outfile}{path of output genetic map file}

\item{vcffile}{a VCF/BCF file with single chromosome ideally}

\item{region}{chromosome name if the VCF has multiple chromosome}
}
\description{
Create a genetic map file from VCF assuming uniform recombination rate
}
