% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatiotemporal_map.R
\name{plot_spatiotemporal_map}
\alias{plot_spatiotemporal_map}
\title{Title: Plotting spatiotemporal maps (time facets) using long sf objects}
\usage{
plot_spatiotemporal_map(long_sf, border_sf, variable_names)
}
\arguments{
\item{long_sf}{A long-format sf object}

\item{border_sf}{Shape files for the borders (eg adm0)}

\item{variable_names}{A vector of variable names that you wish to plot, it will loop over the names if >1}
}
\value{
Description of the value returned by the function.
}
\description{
Plot spatiotemporal maps with time facets based on a vector of variable names to loop over.
This expects a long format sf object (eg covariate data in long format returned from \code{get_map_covar_data()})
and a border shp sf object (this could be the same as the long sf if you want the same borders), alternatively,
you could supply eg admin0 sf to plot only the country borders. Because rendering takes a lot of time, the function
outputs to 'analysis/plots' directory. Will likely edit this to be more felxible (eg user-supplied titles, filenames, color palettes)
}
\examples{
# Example of how to use the function
covars <- readRDS("analysis/data_raw/final_covariates.rds")
border_sf <- readRDS("analysis/data_derived/sf_admin0_africa.rds")
map_data <- get_map_covar_data(covars)
variable_names <- c("pfpr210_mean", "ft", "AL")
plot_spatiotemporal_map(map_data, border_sf, variable_names)
}
