% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{make_STITCH_cli}
\alias{make_STITCH_cli}
\title{Make STITCH command line interface}
\usage{
make_STITCH_cli(function_file, cli_output_file,
  integer_vectors = c("shuffleHaplotypeIterations", "splitReadIterations",
  "refillIterations"), character_vectors = c("reference_populations"),
  other_logical_params = NULL, other_integer_params = NULL,
  other_double_params = NULL, other_character_params = NULL,
  function_name = "STITCH")
}
\arguments{
\item{function_file}{to main STITCH function file}

\item{stitch_cli_file}{where output goes}
}
