% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatiotemporal_map_raster.R
\name{plot_spatiotemporal_map_raster}
\alias{plot_spatiotemporal_map_raster}
\title{Plot spatiotemporal maps from raster data for specified variables}
\usage{
plot_spatiotemporal_map_raster(raster, variable_names)
}
\arguments{
\item{raster}{A raster object (e.g., a SpatRaster from the \strong{terra} package) containing the data to be plotted.}

\item{variable_names}{A vector of variable names corresponding to the layers in the raster object to plot. The function will loop over these names.}
}
\value{
The function does not return a value; it saves plots as PNG files in the 'analysis/plots' directory.
}
\description{
This function plots spatiotemporal maps from a raster object for each variable specified in the vector \code{variable_names}.
It converts the raster data to a data frame, filters out missing values, and creates a plot using \strong{ggplot2} for each variable.
The plots are saved as PNG files in the 'analysis/plots' directory with filenames corresponding to each variable.
Note that the raster object should contain layers corresponding to the variables specified in \code{variable_names}.
}
\examples{
# Example of how to use the function
# library(terra)
# Load your raster data (assuming 'stacked_rasters' is a SpatRaster with
# layers 'prev_Q50', 'prev_Q2.5', 'prev_Q97.5')
# grff_summary <- readRDS("data_derived/GRFF_summary_example.rds")
# sf_grff_summary <- st_as_sf(grff_summary, coords = c("lon", "lat"), crs=4326)
# variable_names <- c("prev_Q50", "prev_Q2.5", "prev_Q97.5")
# plot_spatiotemporal_map_raster(stacked_rasters, variable_names)
}
