% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\alias{stitch.data.table}
\alias{stitch.data.frame}
\title{stitch}
\usage{
stitch(lhs, rhs, on = NULL, ...)

\method{stitch}{data.table}(
  lhs,
  rhs,
  on = NULL,
  lhs.prefix = get_pipe_source_name(lhs),
  rhs.prefix = deparse(substitute(rhs))
)

\method{stitch}{data.frame}(
  lhs,
  rhs,
  on = NULL,
  lhs.prefix = get_pipe_source_name(lhs),
  rhs.prefix = deparse(substitute(rhs))
)
}
\arguments{
\item{lhs}{data; left side data for join}

\item{rhs}{data; right side data for join}

\item{on}{character; join specification see \code{\link[data.table:data.table]{data.table::data.table()}}}

\item{lhs.prefix}{character; name derived from pipe source}

\item{rhs.prefix}{character; prefix for columns from the right}
}
\description{
Safe Left-Join for training/scoring data creation
}
\details{
stitch provides safe left-joins for building rectangularized tables, the
kind used for modeling and other analyses.
\enumerate{
\item Supports LEFT JOINS using \code{on}, \code{key} or a natural join
\item Handles column name collisions by auto prefixing the \code{right-side} data
\item Preserves column ordering.
}
}
\examples{

left <- data.table( letters = letters[1:5], numbers=1:5 )
right <- data.table( lets = letters[1:6], numbers=1:6 )

left \%>\% stitch(right, on=c("letters"="lets" ) )


}
