% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\alias{stitch.data.frame}
\alias{stitch.data.table}
\title{stitch}
\usage{
stitch(lhs, rhs, on = NULL, ...)

\method{stitch}{data.frame}(lhs, rhs, on = NULL, ...)

\method{stitch}{data.table}(lhs, rhs, on = NULL, rhs.prefix = deparse(substitute(rhs)))
}
\arguments{
\item{lhs}{data; left-handed side for LEFT JOIN}

\item{rhs}{data; right-handed side data for LEFT JOIN}

\item{on}{character; join specification see \code{\link[data.table:data.table]{data.table::data.table()}}}

\item{rhs.prefix}{character; prefix for columns from the right}
}
\description{
Safe Left-Join for training/scoring data creation
}
\details{
stitch provides safe LEFT JOINS for building rectangularized tables, the
kind used for modeling and other analyses.
\enumerate{
\item Supports LEFT JOINS using \code{on}, \code{key} or a natural join
\item Preserves column ordering keeping LHS names on the left.
\item Handles column name collisions by auto prefixing the \code{right-side} data
\item Preserves column ordering.
}
}
\examples{

require(data.table)
left <- data.table( id = c(1,1,2,2,4), amount=1:5 )
right <- data.table( id = 1:6, type=letters[1:6]  )

left \%>\% stitch(right)


}
